Derive `Board` and `Folder` from `TileContainer`
