+++ app/Models/Board.php
@@ -3,9 +3,8 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
 
-class Board extends Model
+class Board extends TileContainer
 {
     use HasFactory;
 
@@ -38,34 +37,6 @@ class Board extends Model
         'folders'
     ];
 
-    // append this custom field to the serialized JSON
-    protected $appends = [
-        'contents'
-    ];
-
-    // when this model gets serialized w/ toArray, a field called `contents`
-    // will be populated with the result of this function
-    public function getContentsAttribute()
-    {
-        $contents = $this->folders()->get()->concat($this->words()->get());
-
-        $content_rows = $contents->groupBy(
-            fn ($item) => $item->pivot->board_y
-        )->sortKeys()->values();
-
-        $sorted_rows = $content_rows->map(
-            fn ($row) => $row->sortBy(fn ($item) => $item->pivot->board_x)
-        );
-
-        $expanded_sorted_contents = $sorted_rows->map(function ($row) {
-            return $row->map(function ($item) {
-                return $item->toArray();
-            })->sortKeys()->values();
-        });
-
-        return $expanded_sorted_contents->toArray();
-    }
-
     public function words()
     {
         return $this->belongsToMany(
@@ -93,5 +64,12 @@ class Board extends Model
     public function user()
     {
         return $this->belongsTo(User::class);
+    }
+
+    public function createCopy(): Board
+    {
+        $replicant = $this->replicate();
+
+        return $replicant;
     }
 }
