+++ app/Models/Folder.php
@@ -3,9 +3,8 @@
 namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
 
-class Folder extends Model
+class Folder extends TileContainer
 {
     use HasFactory;
 
@@ -38,11 +37,6 @@ class Folder extends Model
         'folders'
     ];
 
-    // append this custom field to the serialized JSON
-    protected $appends = [
-        'contents'
-    ];
-
     public function words()
     {
         return $this->belongsToMany(
@@ -60,29 +54,6 @@ class Folder extends Model
             'outer_folder_id',
             'inner_folder_id'
         )->withPivot('board_x', 'board_y');
-    }
-
-    // when this model gets serialized w/ toArray, a field called `contents`
-    // will be populated with the result of this function
-    public function getContentsAttribute()
-    {
-        $contents = $this->folders()->get()->concat($this->words()->get());
-
-        $content_rows = $contents->groupBy(
-            fn ($item) => $item->pivot->board_y
-        )->sortKeys()->values();
-
-        $sorted_rows = $content_rows->map(
-            fn ($row) => $row->sortBy(fn ($item) => $item->pivot->board_x)
-        );
-
-        $expanded_sorted_contents = $sorted_rows->map(function ($row) {
-            return $row->map(function ($item) {
-                return $item->toArray();
-            })->sortKeys()->values();
-        });
-
-        return $expanded_sorted_contents->toArray();
     }
 
     public function user()
